refactor BotCallbackService
